+++ go/cwc/io/decoder.go
@@ -15,30 +15,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package cwc
+package io
 
-/**
- * Specifies a device interface interface that can have
- * specific implementations for different I/O setups eg. gpio pins, serial port, audio out generator
- * Implementations exist for Raspberry Pi GPIO and Serial and No IO so far.  See their implementations
- * in this package.
- */
-
-type IO interface {
-	Open() error
-	Bit() bool
-	Dot() bool
-	Dash() bool
-	SetBit(bool)
-	SetToneOut(bool)
-	SetStatusLED(bool)
-	Close()
-}
-
-type ConfigMap map[string]string
-
-// Config consts
-const Keyin = "keyin"
-const Keyout = "keyout"
-const Pcmout = "pcmout"
-const Sidetonefreq = "sidetonefreq"
+// Morse text decoder
+// Uses the generic decoder to output to the console
