+++ go/cwc/io/morse_out.go
@@ -19,7 +19,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 package io
 
 type MorseOut interface {
-	Open() error
+	Open(config *cwc.Config) error
+	ConfigChanged() error
 	SetBit(bool)
 	SetToneOut(bool)
 	SetStatusLED(bool)
