+++ go/cwc/morse_rx.go
@@ -1,0 +1,1 @@
+package cwc
