+++ go/cwc/morse_tx.go
@@ -1,0 +1,1 @@
+package cwc
